Move CaptionGenerator away from v7 package
It is used in LabelField which is a v8 component
