+++ src/main/java/org/vaadin/viritin/fields/LabelField.java
@@ -2,7 +2,6 @@ package org.vaadin.viritin.fields;
 
 import org.vaadin.viritin.fluency.ui.FluentCustomField;
 import org.vaadin.viritin.label.RichText;
-import org.vaadin.viritin.v7.fields.CaptionGenerator;
 
 import com.vaadin.ui.Component;
 import com.vaadin.ui.CustomField;
@@ -14,7 +13,7 @@ import com.vaadin.ui.Label;
  * The use case is a non editable value in a form, e.g. an id.
  * <p>
  * The creation of the label text can be controlled via a
- * {@link org.vaadin.viritin.v7.fields.CaptionGenerator}.
+ * {@link org.vaadin.viritin.fields.CaptionGenerator}.
  *
  * @author Daniel Nordhoff-Vergien
  *
