+++ src/main/java/org/vaadin/viritin/v7/fields/LabelField.java
@@ -3,6 +3,8 @@ package org.vaadin.viritin.v7.fields;
 import com.vaadin.ui.Component;
 import com.vaadin.v7.ui.CustomField;
 import com.vaadin.v7.ui.Label;
+
+import org.vaadin.viritin.fields.CaptionGenerator;
 import org.vaadin.viritin.v7.label.RichText;
 
 /**
@@ -11,7 +13,7 @@ import org.vaadin.viritin.v7.label.RichText;
  * The use case is a non editable value in a form, e.g. an id.
  * <p>
  * The creation of the label text can be controlled via a
- * {@link org.vaadin.viritin.v7.fields.CaptionGenerator}.
+ * {@link org.vaadin.viritin.fields.CaptionGenerator}.
  * 
  * @author Daniel Nordhoff-Vergien
  *
