+++ src/test/java/org/vaadin/viritin/v7/SubSetSelectorExample.java
@@ -6,11 +6,11 @@ import java.util.List;
 import java.util.Set;
 
 import org.vaadin.addonhelpers.AbstractTest;
-import org.vaadin.viritin.v7.fields.CaptionGenerator;
 import org.vaadin.viritin.v7.fields.IntegerField;
 import org.vaadin.viritin.v7.fields.MTextField;
 import org.vaadin.viritin.v7.fields.SubSetSelector;
 import org.vaadin.viritin.v7.form.AbstractForm;
+import org.vaadin.viritin.fields.CaptionGenerator;
 import org.vaadin.viritin.layouts.MFormLayout;
 import org.vaadin.viritin.layouts.MVerticalLayout;
 import org.vaadin.viritin.testdomain.Person;
